add web routes & handlers
